chore: create service function to crank
